All the casts

+++ source/app/layout/circlepackcomponentlayout.cpp
@@ -196,7 +196,8 @@ void CirclePackComponentLayout::executeReal(const Graph& graph, const std::vecto
     for(auto componentId : componentIds)
     {
         auto component = graph.componentById(componentId);
-        float size = (component->numNodes() * 100.0f) / maxNumNodes;
+        float size = (static_cast<float>(component->numNodes()) * 100.0f) /
+            static_cast<float>(maxNumNodes);
         componentLayoutData[componentId].setRadius(size);
     }
